Add Profile to Model
Added Profiles and logically included them in the same places where
Employees are. I tried to limit the amount that "employees" were
completely taken out. I also expanded some Profile functionality,
especially in regards to its relation with User. Specifically, I set the
model up so that Profiles and Users are one-to-one, bi-directional, and
dependent. Also, a User object that is retrieved from a Profile object
is read-only, for security. I also added some helper functions in
Profile that take advantage of the information already housed in User
(such as name, email, and permission level)